currency_code → code

+++ updates/extend_site_definitions_table.php
@@ -9,13 +9,13 @@ return new class extends Migration
     {
         if (!Schema::hasColumn('system_site_definitions', 'currency_id')) {
             Schema::table('system_site_definitions', function(Blueprint $table) {
-                $table->bigInteger('currency_id')->unsigned()->nullable()->index();
+                $table->integer('currency_id')->unsigned()->nullable()->index();
             });
         }
 
         if (!Schema::hasColumn('system_site_definitions', 'base_currency_id')) {
             Schema::table('system_site_definitions', function(Blueprint $table) {
-                $table->bigInteger('base_currency_id')->unsigned()->nullable()->index();
+                $table->integer('base_currency_id')->unsigned()->nullable()->index();
             });
         }
     }
